logging + handle trailing slashes + neaten some things up

+++ manage/api/common.php
@@ -1,21 +1,35 @@
 <?php
     $dataFile = '../data/gogo.txt';
+    $logFile = '../data/logs.txt';
     $separator = '||||';
-            
-    if(!file_exists($dataFile)) {
-        file_put_contents($dataFile,'');
-    }
 
     header('Content-Type: application/json');
+    
+    // create the data/log files if they don't exist
+    if(!file_exists($dataFile)) { file_put_contents($dataFile,''); }
+    if(!file_exists($logFile)) { file_put_contents($logFile,''); }
+
+
+    function logEntry($txt) {
+        global $logFile;
+        $dateTime = date(DATE_ISO8601);
+        file_put_contents($logFile, "[$dateTime]: $txt\n", FILE_APPEND);
+    }
 
     function returnError($error) {
-        echo "{\"error\":\"$error\"}";
-        die();
+        die("{\"error\":\"$error\"}");
     }
 
     function returnSuccess($message) {
-        echo "{\"success\":\"$message\"}";
-        die();
+        die("{\"success\":\"$message\"}");
+    }
+
+    function removeTrailingSlash($input) {
+        if(substr($input, strlen($input)-1, 1) == '/') {
+            return substr($input, 0, -1);
+        } else {
+            return $input;
+        }
     }
 
     function doesEntryExist($source) {
@@ -34,20 +48,21 @@
     function addEntry($source, $dest) {
         global $dataFile, $separator;
         $source = strtolower($source);
+        logEntry("Adding redirect from ($source) to ($dest)");
         file_put_contents($dataFile, "$source$separator$dest\n", FILE_APPEND);
     }
 
     function removeEntry($source) {
         global $dataFile, $separator;
-        $source = strtolower($source);
-
+        $source = removeTrailingSlash(strtolower($source));
+        logEntry("Removing redirect from ($source)");
         $data = explode("\n", file_get_contents($dataFile));
         $i = 0;
         foreach($data as $line) {
             $entry = @explode($separator, $line);
             if(strtolower($entry[0]) == $source) {
-                unset($data[$i]);
-                file_put_contents($dataFile, implode($data,"\n"));
+                unset($data[$i]); // remove the entry at index (line) i
+                file_put_contents($dataFile, implode($data,"\n")); // rewrite them to the file without the removed
             }
             $i++;
         }
